Prevent non-Resource classes from being accessed, Fix debug option.
This patch updates the application exceptions to standardize Resource[Method]Invalid errors, which main.rb now throws when resources are invalid/illegal or can not service a request. The mainTest file now has a block to verify that non-resource classes can not be accessed; only classes that include ResourceInterface are allowed.

Updates the process_request method to correctly print debug information for internalerrors via parameter toggle.