Fix resource identification breaking certain requests
Invalid resource sieving changed the ResourceInterface to a module, which was wrong. Modules do not add class members to classes themselves, but to instances of classes. Consequently, resource classes were unable to access default class http-named methods provided by the interface. This patch updates main.rb to look for Class < Super, which is the correct method to test class inheritance.