Change ResourceInterface to Module (correclty)
The ResourceInterface is now a module, which can be included (into instance resources) and/or extended (into static methods). The main script determines if a requested resource is allowed to handle instance/static requests appropriately.

The mainTest file is updated to verify correct behavior and has also been refactored. This test script necessitated the creation of two demo Resource files (which include xor extend ResourceInterface).